app/requirements.txt
adblockparser==0.7
AdvancedHTTPServer==2.2.0
aiocmd==0.1.2
aioconsole==0.3.1
aiodns==2.0.0
aiohttp==3.7.4
aiomultiprocess==0.8.0
aioredis==1.3.1
aiosqlite==0.16.1
aiowinreg==0.0.3
ajpy==0.0.4
aniso8601==8.1.1
apispec==3.3.1
apispec-webframeworks==0.5.2
appdirs==1.4.4
asciitree==0.3.3
asgiref==3.3.0
asn1crypto==1.4.0
async-timeout==3.0.1
asyncio==3.4.3
asysocks==0.0.2
attrs==20.3.0
autobahn==17.10.1
Automat==20.2.0
Babel==2.8.0
backcall==0.2.0
psycopg2==2.8.6
django
djangorestframework
django-cors-headers
pyjwt
djangorestframework-jwt
easy-thumbnails